requirements.txt
anyio==3.3.4
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
joblib==1.1.0
numpy==1.21.4
pandas==1.3.4
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
threadpoolctl==3.0.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
